# Speed up homepage by calculating "interesting" things more efficiently