The homepage is currently really slow to load - nearly 2s on my machine. A lot of this is due to the inefficient way in which we're calculating the lists of interesting crops and plantings - we load everything into memory from the DB, and then loop over them in Ruby, emitting more queries for each object to find out if it's interesting. With some moderately dark ActiveRecord magic we can do all of this calculation in only two database queries, getting page load time down to around 0.8s - still not great, but much better than it was.

Before
---------
```
$ ab -n 20 http://127.0.0.1:8080/
This is ApacheBench, Version 2.3 <$Revision: 1638069 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking 127.0.0.1 (be patient).....done


Server Software:________
Server Hostname:        127.0.0.1
Server Port:            8080

Document Path:          /
Document Length:        23975 bytes

Concurrency Level:      1
Time taken for tests:   37.634 seconds
Complete requests:      20
Failed requests:        0
Total transferred:      495740 bytes 
HTML transferred:       479500 bytes 
Requests per second:    0.53 [#/sec] (mean) 
Time per request:       1881.699 [ms] (mean)
Time per request:       1881.699 [ms] (mean, across all concurrent requests)
Transfer rate:          12.86 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.0      0       0
Processing:  1832 1882  49.9   1872    1991
Waiting:     1832 1881  49.9   1872    1990
Total:       1832 1882  49.9   1872    1991
  
Percentage of the requests served within a certain time (ms)
  50%   1872
  66%   1883
  75%   1901
  80%   1955
  90%   1978
  95%   1991
  98%   1991 
  99%   1991
 100%   1991 (longest request)
```

After
-------
```
$ ab -n 20 http://127.0.0.1:8080/
This is ApacheBench, Version 2.3 <$Revision: 1638069 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking 127.0.0.1 (be patient).....done


Server Software:________
Server Hostname:        127.0.0.1
Server Port:            8080

Document Path:          /
Document Length:        28299 bytes

Concurrency Level:      1
Time taken for tests:   15.677 seconds
Complete requests:      20
Failed requests:        0
Total transferred:      582220 bytes
HTML transferred:       565980 bytes
Requests per second:    1.28 [#/sec] (mean)
Time per request:       783.854 [ms] (mean)
Time per request:       783.854 [ms] (mean, across all concurrent requests)
Transfer rate:          36.27 [Kbytes/sec] received

Connection Times (ms)   
              min  mean[+/-sd] median   max
Connect:        0    0   0.0      0       0
Processing:   768  784  11.7    785     815
Waiting:      767  783  11.7    784     814
Total:        768  784  11.7    785     815

Percentage of the requests served within a certain time (ms)
  50%    785
  66%    788
  75%    791
  80%    793
  90%    798
  95%    815
  98%    815
  99%    815
 100%    815 (longest request)
```