@pozorvlak 2015-07-11
I tried using the `plantings_count` field for `Crop.has_plantings`, but it failed the tests for interestingness - I'm not sure why. Further speedups would require an actual profiler, and I'm having great difficulty finding instructions for profiling Rails apps that aren't out-of-date: can anyone help?