@cesy 2015-07-17
Can we cache the results, and just refresh it every time there's a photo or planting change? For something like this, sensible caching will make an even bigger difference.