@cesy 2015-07-17
The changes look sensible, but do we have any tests covering that it returns the right thing? It looks like our view spec on these isn't very thorough, there isn't a feature test, and the model spec is extremely light.