@maco 2015-07-17
Right now we have caching of posts on the front page and it works
*horrendously*.

On Fri, Jul 17, 2015 at 5:50 AM Cesy <notifications@github.com> wrote:

> The changes look sensible, but do we have any tests covering that it
> returns the right thing? It looks like our view spec on these isn't very
> thorough, there isn't a feature test, and the model spec is extremely light.
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/Growstuff/growstuff/pull/775#issuecomment-122231833>.
>