@pozorvlak 2015-07-17
The `cache cache_key_for` lines in eg `app/views/home/_crops.html.haml` are meant to cache this stuff already, but I still get a depressing email from NewRelic every Monday about how slow our page-load times are :-(