@cesy 2015-07-20
Sounds like something's not working properly, then. Can you add me to
NewRelic and I'll prod a bit more? In the meantime, this is almost
certainly an improvement, but I'd be more comfortable adding some tests on
it before merging.

On 17 July 2015 at 17:01, pozorvlak <notifications@github.com> wrote:

> The cache cache_key_for lines in eg app/views/home/_crops.html.haml are
> meant to cache this stuff already, but I still get a depressing email from
> NewRelic every Monday about how slow our page-load times are :-(
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/Growstuff/growstuff/pull/775#issuecomment-122326983>.
>