@pozorvlak 2015-07-21
@cesy: I can't see how to add you to the NewRelic email alerts, but if you go to https://dashboard.heroku.com/apps/growstuff-prod/resources you should be able to click through to the NewRelic dashboard. Actually, try this link: https://rpm.newrelic.com/accounts/294339/applications.