@cesy 2015-08-05
Okay, apparently I was looking at the wrong model spec. The planting one is pretty detailed and tests interesting things thoroughly enough that I'm happy to merge this. And NewRelic works, so I'll start having a look at other bits next.

@maco what's wrong with the caching of posts on the front page?