@cesy 2015-08-06
Fixed! :)

On 5 August 2015 at 23:41, Mackenzie <notifications@github.com> wrote:

> What's wrong on the front page is it showing things from weeks ago as 1
> minute ago
> On Wed, Aug 5, 2015 at 2:19 PM Cesy <notifications@github.com> wrote:
>
> > Merged #775 <https://github.com/Growstuff/growstuff/pull/775>.
> >
> > —
> > Reply to this email directly or view it on GitHub
> > <https://github.com/Growstuff/growstuff/pull/775#event-374504970>.
>
> >
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/Growstuff/growstuff/pull/775#issuecomment-128172401>.
>